Completed the header
I have added a logo (my name with a title underneath) but outside the navbar as it would interfere with the nav items due to the bootstrap styling, and positioned it absolute to be inline with the nav items. I made an active class for the nav items and added it to their respective pages. I added a font to the body, logo, and nav items. I made media queries to get a perfect header across all screen sizes.