Added a footer and font awesome link
I have added a basic footer with 4 social media links with font awesome icons. The facebook and twitter links open up to the login pages only but the github and linked in links go to my actual profiles. All in a new tab. With bootstrap classes and have centralised the footer and made it 100% width across all devices.