Styled the font awesome icons and corrected bug in header
I made all the icons circular and added a hover effect which increases the scale. I also added that effect to the nave icons in header. The logo <a> tag was under the navber so it wasn't accessible. Adding z-index fixed this. I removed border from the header as it appeard to contradict the footer style.