Added content to About me page
I have added another profile picture of myself. A funny one to be ideal as the title is About Me. I have also decided to remove the background image on the index page as it slowed performance and looked inconsistent with the About Me page. I removed the margin on the row divs as it was causing sidescroll. I also added a welcome message below my profile picture in the index page to be consistent with the About Me page. This caused me to change the header tags respectively for better accessability.