Added an interests section to the about me page
This is the final section to the about me page, with a white background to show a new section. It has been styled respectively to the rest of the pages headings and lists except I've chosen to have the lsit style as bullet points and not have the text centered as it has more writing and it is easier to read this way. I only need to add an icon with a prompt to download CV and then I'll be moving on to the contact page.