added a pdf of my CV to download
I have a heading 'Download my CV here' pointing to a download icon. All has been styled respectively. I have added the hr tag above it which only appears when screen width is medium. This is for when the interests lists go into three columns so it seperates it from that content