Completed the contact me page with a single form
This page simply has just the
form for contacting me. It has been styled respectively. The unique style I have given it is the background colour which is the navy gray throughout the project, white border, and solid box shadow which is the same navy gray giving an inset white border effect. The main thing about the form is the use of formspree. This is a way of sending an email with no javascript and keeps my email safe and protects me from spam