Completed the content of the portfolio page (Portfolio heading and cards)
I took the previous html page's structure and added one section for the content of the portfolio. I made one container for the heading and text, and then another row beneath containing all the portfolio cards which will display screen shots of future projects which can also be clicked on as a link to the live website. Each card has a button which holds the bootstrap accordion dropdown function which will display information regarding the project and contain a link to the github repositry.